epee/portable_storage: move code to .cpp
There's a bunch of code in inline functions in the header for no good
reason.  Move it to a .cpp.
