+++ contrib/epee/src/portable_storage.cpp
@@ -1,4 +1,5 @@
 #include "storages/portable_storage_to_json.h"
+#include "storages/portable_storage.h"
 
 namespace epee {
   namespace serialization {
@@ -79,5 +80,123 @@ namespace epee {
       }
     }
 
+    bool portable_storage::dump_as_json(std::string& buff, size_t indent, bool insert_newlines)
+    {
+      TRY_ENTRY();
+      std::stringstream ss;
+      epee::serialization::dump_as_json(ss, m_root, indent, insert_newlines);
+      buff = ss.str();
+      return true;
+      CATCH_ENTRY("portable_storage::dump_as_json", false)
+    }
+
+    bool portable_storage::load_from_json(std::string_view source)
+    {
+      TRY_ENTRY();
+      return json::load_from_json(source, *this);
+      CATCH_ENTRY("portable_storage::load_from_json", false)
+    }
+
+    bool portable_storage::store_to_binary(std::string& target)
+    {
+      TRY_ENTRY();
+      std::stringstream ss;
+      storage_block_header sbh{};
+      sbh.m_signature_a = PORTABLE_STORAGE_SIGNATUREA;
+      sbh.m_signature_b = PORTABLE_STORAGE_SIGNATUREB;
+      sbh.m_ver = PORTABLE_STORAGE_FORMAT_VER;
+      ss.write(reinterpret_cast<const char*>(&sbh), sizeof(storage_block_header));
+      pack_entry_to_buff(ss, m_root);
+      target = ss.str();
+      return true;
+      CATCH_ENTRY("portable_storage::store_to_binary", false)
+    }
+
+    bool portable_storage::load_from_binary(const epee::span<const uint8_t> source)
+    {
+      m_root.m_entries.clear();
+      if(source.size() < sizeof(storage_block_header))
+      {
+        LOG_ERROR("portable_storage: wrong binary format, packet size = " << source.size() << " less than expected sizeof(storage_block_header)=" << sizeof(storage_block_header));
+        return false;
+      }
+      storage_block_header* pbuff = (storage_block_header*)source.data();
+      if(pbuff->m_signature_a != PORTABLE_STORAGE_SIGNATUREA ||
+        pbuff->m_signature_b != PORTABLE_STORAGE_SIGNATUREB
+        )
+      {
+        LOG_ERROR("portable_storage: wrong binary format - signature mismatch");
+        return false;
+      }
+      if(pbuff->m_ver != PORTABLE_STORAGE_FORMAT_VER)
+      {
+        LOG_ERROR("portable_storage: wrong binary format - unknown format ver = " << pbuff->m_ver);
+        return false;
+      }
+      TRY_ENTRY();
+      throwable_buffer_reader buf_reader(source.data()+sizeof(storage_block_header), source.size()-sizeof(storage_block_header));
+      buf_reader.read(m_root);
+      return true;//TODO:
+      CATCH_ENTRY("portable_storage::load_from_binary", false);
+    }
+
+    section* portable_storage::open_section(const std::string& section_name, section* parent_section, bool create_if_notexist)
+    {
+      TRY_ENTRY();
+      if (!parent_section) parent_section = &m_root;
+
+      storage_entry* pentry = find_storage_entry(section_name, parent_section);
+      if(!pentry)
+      {
+        if(!create_if_notexist)
+          return nullptr;
+        return insert_new_section(section_name, parent_section);
+      }
+      CHECK_AND_ASSERT(pentry , nullptr);
+      //check that section_entry we find is real "CSSection"
+      if (!std::holds_alternative<section>(*pentry))
+      {
+        if(create_if_notexist)
+          *pentry = section();//replace
+        else
+          return nullptr;
+      }
+      return &std::get<section>(*pentry);
+      CATCH_ENTRY("portable_storage::open_section", nullptr);
+    }
+
+    bool portable_storage::get_value(const std::string& value_name, storage_entry& val, section* parent_section)
+    {
+      //TRY_ENTRY();
+      if(!parent_section) parent_section = &m_root;
+      storage_entry* pentry = find_storage_entry(value_name, parent_section);
+      if(!pentry)
+        return false;
+
+      val = *pentry;
+      return true;
+      //CATCH_ENTRY("portable_storage::template<>get_value", false);
+    }
+
+    storage_entry* portable_storage::find_storage_entry(const std::string& pentry_name, section* psection)
+    {
+      TRY_ENTRY();
+      CHECK_AND_ASSERT(psection, nullptr);
+      auto it = psection->m_entries.find(pentry_name);
+      if(it == psection->m_entries.end())
+        return nullptr;
+
+      return &it->second;
+      CATCH_ENTRY("portable_storage::find_storage_entry", nullptr);
+    }
+
+    section* portable_storage::insert_new_section(const std::string& pentry_name, section* psection)
+    {
+      TRY_ENTRY();
+      storage_entry* pse = insert_new_entry_get_storage_entry(pentry_name, psection, section());
+      if(!pse) return nullptr;
+      return &std::get<section>(*pse);
+      CATCH_ENTRY("portable_storage::insert_new_section", nullptr);
+    }
   }
 }
